style : page update activity
