+++ src/app/modules/activity/components/feature/activity-update-management/activity-update-management.component.html
@@ -1,0 +1,108 @@
+<app-activity-form-layout
+	pageTitle="Modifier une activité"
+	backgroundImage="/assets/photos/update-activity.jpg"
+>
+	<form #updateActivityForm="ngForm" (ngSubmit)="onSubmit(updateActivityForm)">
+		<app-avatar-pseudo></app-avatar-pseudo>
+
+		<app-text-field
+			labelFor="name"
+			labelContent="Titre de l'activité :"
+			type="text"
+			id="name"
+			name="name"
+			[form]="updateActivityForm"
+			[(ngModel)]="formData.name"
+			[required]="true"
+		></app-text-field>
+
+		<app-input-date
+			labelFor="date"
+			labelContent="Date de l'activité :"
+			inputId="date"
+			id="date"
+			name="date"
+			[showIcon]="true"
+			#date="ngModel"
+			[(ngModel)]="formData.date"
+		></app-input-date>
+
+		<app-input-hour
+			labelFor="date"
+			labelContent="Heure de l'activité :"
+			inputId="hour"
+			id="hour"
+			name="hour"
+			#hour="ngModel"
+			[(ngModel)]="formData.hour"
+		></app-input-hour>
+
+		<app-select-city
+			labelFor="city"
+			labelContent="Lieu :"
+			id="city"
+			inputId="city"
+			name="activityCity"
+			#activityCity="ngModel"
+			[(ngModel)]="formData.activityCity"
+		>
+		</app-select-city>
+
+		<app-select-category
+			labelFor="categoryId"
+			labelContent="Catégorie :"
+			id="categoryId"
+			inputId="categoryId"
+			name="categoryId"
+			#category="ngModel"
+			[(ngModel)]="formData.categoryId"
+		></app-select-category>
+
+		<app-select-number
+			labelFor="nbGuest"
+			labelContent="Nbre de participants :"
+			inputId="nbGuest"
+			name="nbGuest"
+			#nbGuest="ngModel"
+			[(ngModel)]="formData.nbGuest"
+		></app-select-number>
+
+		<app-textarea
+			labelFor="description"
+			labelContent="Description de l'évènement :"
+			id="description"
+			name="description"
+			[cols]="5"
+			[rows]="5"
+			#description="ngModel"
+			[form]="updateActivityForm"
+			[(ngModel)]="formData.description"
+		></app-textarea>
+
+		<p>img</p>
+		<i class="pi pi-download" style="font-size: 1.5rem"></i>
+
+		<input
+			inputId="imgUrl"
+			name="imgUrl"
+			#imgUrl="ngModel"
+			[(ngModel)]="formData.imgUrl"
+			type="file"
+			class="file-upload"
+		/>
+
+		<app-text-field
+			labelFor="link"
+			labelContent="Lien :"
+			type="url"
+			id="link"
+			name="link"
+			[form]="updateActivityForm"
+			[(ngModel)]="formData.link"
+		></app-text-field>
+
+		<div>
+			<p-button rounded label="Modifier" type="submit"></p-button>
+		</div>
+	</form>
+</app-activity-form-layout>
