+++ src/app/modules/activity/components/feature/activity-update-management/activity-update-management.component.scss
